requirements.txt
Flask==2.2.5
tensorflow==2.10.0
pillow==9.2.0
numpy==1.23.1
